smartsheet_project/pdf-to-csv-func/requirements.txt
azure-functions
camelot-py[cv]
ghostscript
opencv-python
